requirements.txt

Django==4.1.5
djangorestframework==3.14.0
Pillow==9.4.0
python-dotenv==0.21.0
djoser==2.1.0
drf-yasg==1.21.4

